// src/modules/tariff/types.ts
export interface Tariffs {
  tariff_id: number;
  title: string;
  tariff_typing_id: number;
  description: string;
  price: number;
  deadline: number;
  sms: string;
  minutes: string;
  mega_byte: string;
  sale: any;
  status: number;
  ussd_code: string;
  company_name: string;
}